refactor(mvc): cleanse package structure
