+++ mvc/src/main/java/nextstep/mvc/adapter/ModelAndViewHandlerAdapter.java
@@ -1,19 +1,16 @@
-package nextstep.mvc.controller.tobe.adapter;
+package nextstep.mvc.adapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.controller.tobe.HandlerExecution;
+import nextstep.mvc.mapping.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
 
 public class ModelAndViewHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(Object handler) {
-        if (handler instanceof HandlerExecution) {
-            return ((HandlerExecution) handler).hasReturnTypeOf(ModelAndView.class);
-        }
-        return false;
+        return ((HandlerExecution) handler).hasReturnTypeOf(ModelAndView.class);
     }
 
     @Override
