+++ mvc/src/main/java/nextstep/mvc/mapping/AnnotationHandlerMapping.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.mapping;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.lang.reflect.Method;
