+++ mvc/src/main/java/nextstep/mvc/mapping/HandlerKey.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.mapping;
 
 import nextstep.web.support.RequestMethod;
